fix tests and move jest config to package
