+++ .gitignore
@@ -1,8 +1,6 @@
 node_modules
 coverage
-.DS_Store
 
+.DS_Store
 *.js
 *.d.ts
-
-!*.config.js
