--- jest.config.ts
@@ -1,8 +1,0 @@
-import type { Config } from '@jest/types';
-
-export default async (): Promise<Config.InitialOptions> => ({
-	transform: { '^.+\\.ts?$': 'ts-jest' },
-	testEnvironment: 'node',
-	testRegex: '/tests/.*\\.(test|spec)?\\.(ts|tsx)$',
-	moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-});
